+++ src/_column_offsets.ts
@@ -1,7 +1,7 @@
 import { reduce } from '@ctx-core/array'
-import type { column_offsets_T } from './column_offsets_T'
-export function _column_offsets(columns:string[]):Record<string, number> {
-	return reduce<string, column_offsets_T>(
+import type { column_offsets_I } from './column_offsets_I'
+export function _column_offsets(columns:string[]):column_offsets_I {
+	return reduce<string, column_offsets_I>(
 		columns, (memo, column, i
 		)=>{
 			memo[column] = i