+++ src/_rows.ts
@@ -1,10 +1,10 @@
 import { _row_proxy } from './_row_proxy'
 import type { $table_T } from './table_b'
-import type { column_offsets_T } from './column_offsets_T'
+import type { column_offsets_I } from './column_offsets_I'
 import type { Row } from './Row'
 export function _rows<I extends unknown, O extends Row = Row>(
 	table:$table_T<I>,
-	column_offsets:column_offsets_T
+	column_offsets:column_offsets_I
 ) {
 	if (!table || !column_offsets) return
 	const data_row_a1 = table.slice(1) as I[][]