+++ src/column_offsets_I.ts
@@ -1,0 +1,1 @@
+export type column_offsets_I = Record<string, number>