+++ src/column_offsets_T.ts
@@ -1,1 +1,0 @@
-export interface column_offsets_T extends Record<string, number> {}